Extract pronoun helpers into modules
